not needed cast removed

+++ android/src/main/java/com/stellarscript/vlcvideo/VLCVideoViewManager.java
@@ -111,7 +111,7 @@ final class VLCVideoViewManager extends SimpleViewManager<VLCVideoView> {
                         args.size() > 0 &&
                         !args.isNull(0) &&
                         args.getType(0) == ReadableType.Number) {
-                    final int id = (int) args.getInt(0);
+                    final int id = args.getInt(0);
                     videoView.setSubtitlesTrack(id);
                 }
                 break;
@@ -120,7 +120,7 @@ final class VLCVideoViewManager extends SimpleViewManager<VLCVideoView> {
                         args.size() > 0 &&
                         !args.isNull(0) &&
                         args.getType(0) == ReadableType.Number) {
-                    final int id = (int) args.getInt(0);
+                    final int id = args.getInt(0);
                     videoView.setAudioTrack(id);
                 }
                 break;
